gateway/requirements.txt
Flask==3.0.0
requests==2.32.0
PyJWT==2.9.0
